Experimental installer, based on r33 of the trunk
This is a very basic web based installer and should be considered 
experimental.

It tried to detect if Timesheet NG is installed, if not, takes user to 
the installer. If Timesheet NG is installed, but of a lower version, it 
takes user to an (as yet non-working) upgrade page.

The timesheet.sql.in file is now deprecated, the installation SQL being 
moved to install/index.php, as is the install.sh file.

To install a clean copy of Timesheet NG, simply checkout this version 
from SVN into a directory within your web server, navigate in your 
browser to that directory and you should be automatically redirected to 
the web based installation. This page will have instructions to proceed.

Todo:
1. Get the "Check Database Connection" ajax functionality working
2. Allow for the creation of the database
3. Allow for the upgrading from previous versions
4. Tidy up now unused *.inc.in files