I made a series of changes. Most of them have been added to the bug tracker as attached files previously. The changes are as follows:
A dropdown list for filtering projects by status in 
the proj_maint.php page.

I separated the client/project dropdown list to two 
drop downs, one for filtering clients and one for 
filtering projects.
This is less confusing when the user is assigned to 
a lot of projects.

I also commented out the A_DAY constant from the 
simple_action.php, since it is now declared in the 
common.inc file.

I fixed a few bugs in report_specific_client.php 
regarding the print functionality when in weekly 
mode. Navigating to a specific week and pressing the 
print button would cause the page viewed in 
report_specific_client.php to change. This was due 
to a missing hidden field in for "day" in the form. 
The print preview didn't show the data correct data. 
The problem was caused by a missing mode $_GET 
parameter in $print_popup_href string.

To my knowledge the following feature existed in an 
earlier version.
I’ve added a report option, that lets the user 
generate a report for the hours recorded for a 
specific client. I’ve added both the printing and 
Excel functionality.

Users, who are associated with very large amounts of 
data, often face problems when using pages like 
Projects. Projects prints all the data specified by 
the filtering selection lists on one long page and 
when the user for instance filters for "all clients" 
and "all statuses" (or any other large query), it 
can even cause performance problems. These long 
pages are both confusing and slow to load.

I solved this by implementing page separation. Now, 
a limited number of projects are printed, on 
multiple pages. The user can navigate between these 
paging with this menu:

Previous Page / Next Page ( 1 of 797 )

The number of items that is printed per page can be 
specified in “Configuration” under “Items per page 
in Projects”. The default number of items is 10.

I also have added page separation to Tasks.

Finally, I changed the background color of the even 
rows in a lot of tables, to increase their clarity. 
This includes Daily Timesheet, Clients, 
Project-rates, Reports, Users and User-rates.