Editing an entry via the daily form subsequently caused the daily form to only display that project's times with no indication that the displayed entries were filtered, and no way of changing that filter on the page. Changing from there to the simple form carried this same problem to the simple form as well.
Changed the simple and daily forms to always display all entries for a 
given date, and changed the edit.php form to include the client_id, so 
that client, project, task could be displayed as the "default" entry in 
the clock on/off form.  This does have the potentially undesirable effect 
of setting the client/project/task id's for other forms as well, but the
other forms have the ability to change those items.