Fixed several small things that caused errors in the html error logs. Fixed bug that caused fixStartEndDuration to not fix anything. Cleaned up assign-proj-mbrs-to-all-tasks warning message.
Added ability to administratively close the site by creating a file
with the name "siteclosed" in the main context directory.  To reopen
simply delete that file.

Added a favicon file, you'll need to add a line to the headerhtml
configuration item that reads:
  <link rel="shortcut icon" href="images/favicon.ico">
in order for it to start showing up.